DEV-7511: AverageGetTime now kept in nanos DEV-7554: added CacheHitRatio as sampled stat with history
git-svn-id: http://svn.terracotta.org/svn/ehcache@5872 b9324663-ca0f-0410-8574-be9b3887307d

+++ branches/ehcache-core-2.6.x/src/main/java/net/sf/ehcache/management/sampled/CacheSampler.java
@@ -492,7 +492,7 @@ public interface CacheSampler extends LiveCacheStatistics, SampledCacheStatistic
     long getCacheExpirationRate();
 
     /**
-     * @return average get time (ms.)
+     * @return average get time (nanos.)
      */
-    float getCacheAverageGetTime();
+    long getCacheAverageGetTime();
 }
